Revert first pass at refactoring
New refactoring is in the plotting/new submodule
